+++ cli/build.rs
@@ -2,6 +2,7 @@
 
 use std::env;
 use std::path::PathBuf;
+use std::sync::Arc;
 
 use deno_core::snapshot_util::*;
 use deno_core::Extension;
@@ -361,7 +362,10 @@ fn create_cli_snapshot(snapshot_path: PathBuf) {
     deno_http::deno_http::init_ops(),
     deno_io::deno_io::init_ops(Default::default()),
     deno_fs::deno_fs::init_ops::<_, PermissionsContainer>(false, StdFs),
-    deno_node::deno_node::init_ops::<deno_runtime::RuntimeNodeEnv>(None),
+    deno_node::deno_node::init_ops::<deno_runtime::RuntimeNodeEnv>(
+      None,
+      Some(Arc::new(deno_node::RealFs)),
+    ),
     cli::init_ops_and_esm(), // NOTE: This needs to be init_ops_and_esm!
   ];
