+++ cli/tools/check.rs
@@ -9,6 +9,7 @@ use deno_core::error::AnyError;
 use deno_graph::Module;
 use deno_graph::ModuleGraph;
 use deno_runtime::colors;
+use deno_runtime::deno_node::NodeResolver;
 use once_cell::sync::Lazy;
 use regex::Regex;
 
@@ -21,7 +22,6 @@ use crate::cache::Caches;
 use crate::cache::DenoDir;
 use crate::cache::FastInsecureHasher;
 use crate::cache::TypeCheckCache;
-use crate::node::CliNodeResolver;
 use crate::npm::CliNpmResolver;
 use crate::tsc;
 use crate::version;
@@ -42,7 +42,7 @@ pub struct TypeChecker {
   deno_dir: DenoDir,
   caches: Arc<Caches>,
   cli_options: Arc<CliOptions>,
-  node_resolver: Arc<CliNodeResolver>,
+  node_resolver: Arc<NodeResolver>,
   npm_resolver: Arc<CliNpmResolver>,
 }
 
@@ -51,7 +51,7 @@ impl TypeChecker {
     deno_dir: DenoDir,
     caches: Arc<Caches>,
     cli_options: Arc<CliOptions>,
-    node_resolver: Arc<CliNodeResolver>,
+    node_resolver: Arc<NodeResolver>,
     npm_resolver: Arc<CliNpmResolver>,
   ) -> Self {
     Self {
