+++ runtime/build.rs
@@ -222,7 +222,6 @@ mod startup_snapshot {
 
   impl deno_node::NodeEnv for SnapshotNodeEnv {
     type P = Permissions;
-    type Fs = deno_node::RealFs;
   }
 
   deno_core::extension!(runtime,
@@ -324,7 +323,7 @@ mod startup_snapshot {
       runtime::init_ops_and_esm(),
       // FIXME(bartlomieju): these extensions are specified last, because they
       // depend on `runtime`, even though it should be other way around
-      deno_node::deno_node::init_ops_and_esm::<SnapshotNodeEnv>(None),
+      deno_node::deno_node::init_ops_and_esm::<SnapshotNodeEnv>(None, None),
       #[cfg(not(feature = "snapshot_from_snapshot"))]
       runtime_main::init_ops_and_esm(),
     ];
