+++ runtime/lib.rs
@@ -39,5 +39,4 @@ pub use worker_bootstrap::BootstrapOptions;
 pub struct RuntimeNodeEnv;
 impl deno_node::NodeEnv for RuntimeNodeEnv {
   type P = permissions::PermissionsContainer;
-  type Fs = deno_node::RealFs;
 }
